[CHANGE] load athlete from db

+++ app/src/main/java/de/ironcoding/fitsim/logic/Exercise.java
@@ -8,9 +8,9 @@ public class Exercise extends Activity {
 
     private static final int TYPE_INT = 0;
 
-    private final int strainedMuscle;
+    private final long strainedMuscle;
 
-    public Exercise(String name, float pal, int effort, int experience, int strainedMuscle, int minLevel) {
+    public Exercise(String name, float pal, int effort, int experience, long strainedMuscle, int minLevel) {
         super(name, pal, effort, experience, minLevel, TYPE_INT);
         this.strainedMuscle = strainedMuscle;
     }
